requirement.txt
asgiref==3.5.2
auditlog3==1.0.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.32.0
cryptography==37.0.4
cssselect2==0.7.0
dj-database-url==1.2.0
Django==4.1
django-autoslug==1.9.8
django-bootstrap-daterangepicker==1.0.6
django-calculation==1.0.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-countries==7.5.1
django-filter==22.1
django-heroku==0.3.1
django-hvad==1.8.0
django-mathfilters==1.0.0
django-model-mixins==0.1.3
django-multiselectfield==0.1.12
django-nested-admin==4.0.2
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-use-email-as-username==1.2.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
fonttools==4.39.0
html5lib==1.1
idna==3.4
inflection==0.5.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
Pillow==9.2.0
protobuf==3.20.1
psycopg2==2.9.5
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyphen==0.13.2
python-dateutil==2.8.2
python-decouple==3.6
python-monkey-business==1.0.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.2.1
twilio==7.15.4
typing_extensions==4.5.0
tzdata==2022.2
urllib3==1.26.13
weasyprint==58.1
webencodings==0.5.1
whitenoise==6.4.0
zopfli==0.2.2